Football Management System v1b completed

+++ src/main/java/com/projects/Football/Club/Management/System/service/TeamService.java
@@ -38,7 +38,10 @@ public class TeamService {
     public void updateTeam( Team team){
         repo.save(team);
     }
-    public void deletePlayer(int teamId){
+
+    public void deleteTeam(int teamId){
+        if(repo.existsById(teamId))
+            return;
         repo.deleteById(teamId);
     }
 
@@ -47,10 +50,45 @@ public class TeamService {
         Player player = playerRepo.findById(playerId).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
         if(team.getPlayers().size() < maxSquadSize) {
             if (player.getTeam() == null) {
+                if(playerRepo.existsByTeamIdAndJerseyNumber(teamId,player.getJerseyNumber()))
+                    return;
                 player.setTeam(team);
                 team.getPlayers().add(player);
                 playerRepo.save(player);
             }
         }
+    }
+
+    public void removePlayerFromTeam(int teamId, int playerId) {
+        Team team = repo.findById(teamId).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        Player player = playerRepo.findById(playerId).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        if(player.getTeam() != null){
+        if(player.getTeam().getId().equals(team.getId())){
+            team.getPlayers().remove(player);
+            player.setTeam(null);
+            playerRepo.save(player);
+        }
+        }
+    }
+
+    public void transferPlayer(int teamId1, int teamId2, int playerId) {
+        Team team1 = repo.findById(teamId1).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        Team team2 = repo.findById(teamId2).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        Player player = playerRepo.findById(playerId).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+        if(player.getTeam() == null)
+            return;
+        if(!player.getTeam().getId().equals(team1.getId()))
+            return;
+        if(team1.getId() == team2.getId())
+            return;
+        if(!(team2.getPlayers().size() < maxSquadSize))
+        return;
+        if(playerRepo.existsByTeamIdAndJerseyNumber(teamId2,player.getJerseyNumber()))
+            return;
+
+        team1.getPlayers().remove(player);
+        team2.getPlayers().add(player);
+        player.setTeam(team2);
+        playerRepo.save(player);
     }
 }
